+++ Plutarch/Rational.hs
@@ -16,7 +16,7 @@ module Plutarch.Rational (
 import Data.Kind (Type)
 import Data.Ratio (denominator, numerator)
 import GHC.Generics (Generic)
-import Plutarch.Bool (PEq, POrd, PPartialOrd, pif, pnot, (#<), (#<=), (#==))
+import Plutarch.Bool (POrd, PPartialOrd, pif, pnot, (#<), (#<=))
 import Plutarch.Builtin (
   PAsData,
   PBuiltinList,
@@ -30,6 +30,7 @@ import Plutarch.Builtin (
   pfromDataImpl,
  )
 import Plutarch.Integer (PInteger, PIntegral (pquot), pdiv, pmod)
+import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (pconstant)
 import Plutarch.Internal.Other (pfix, pto)
 import Plutarch.Internal.PLam (plam)